% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_row.R
\name{add_row}
\alias{add_row}
\title{Add a row to a selected table}
\usage{
add_row(table, field, value)
}
\arguments{
\item{table}{a name of the modified table,}

\item{field}{names of columns,}

\item{value}{inserted values.}
}
\description{
The function adds a row to a selected table.
Cells which are not provided by a user are replaced with \code{NA}.
}
