% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_row.R
\name{delete_row}
\alias{delete_row}
\title{Delete a row from a selected table}
\usage{
delete_row(table, row_key)
}
\arguments{
\item{table}{a name of the modified table,}

\item{row_key}{a unique name identifing a row.}
}
\description{
The function deletes a row from a selected table.
}
