% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_load.R
\docType{data}
\name{geo_data}
\alias{geo_data}
\title{Environment for geographic data}
\format{
An object of class \code{environment} of length 0.
}
\usage{
geo_data
}
\description{
The environment is created automatically within the package
environment.
}
\keyword{datasets}
