% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{geography}
\alias{geography}
\title{Geography database}
\format{
Data frames:
\describe{
\item{city}{contain information about cities around the world,}
\item{country}{contain information about countries,}
\item{country_language}{contain information about languages used in countries.}
}
}
\description{
A very simple database of geographic information.
}
