% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_tab.R
\name{join_tab}
\alias{join_tab}
\title{Join selected tables}
\usage{
join_tab(table_1, table_2, key_col_1, key_col_2, sel_fun = NULL)
}
\arguments{
\item{table_1}{a name of the first table,}

\item{table_2}{a name of the second table,}

\item{key_col_1}{a column with unique names in the first table,}

\item{key_col_2}{a column with unique names in the second table,}

\item{sel_fun}{a function selecting rows according to a user defined criterion.}
}
\value{
a resulting data frame
}
\description{
The function joins (SQL: inner join) selected tables
with respect to a key.
}
