% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load geographic database}
\usage{
load_data()
}
\description{
The function loads three data frames (city, country and
country language) to a new environment named \code{geography_data}.
It allows package functions to work on data frames without copying them.
}
