% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tab.R
\name{read_tab}
\alias{read_tab}
\title{Read contents of a selected table}
\usage{
read_tab(table, field, sel_fun = NULL, ...)
}
\arguments{
\item{table}{a name of a table to read from,}

\item{field}{names of columns to print,}

\item{sel_fun}{a function selecting rows according to a user defined criterion.}

\item{...}{argument passed to \code{print} function}
}
\description{
The function reads contents of a selected table and
prints it to the console.
}
