% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{sel_row}
\alias{sel_row}
\title{Create a function selecting rows}
\usage{
sel_row(table, ...)
}
\arguments{
\item{table}{a name of a table,}

\item{...}{arguments of character strings type, describing conditions.}
}
\description{
The function uses conditions provided by a user to create
a selection function which role is to select rows printed by
\code{read_tab} function.
}
