% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_row.R
\name{update_row}
\alias{update_row}
\title{Update a row in a selected table}
\usage{
update_row(table, row_key, field, value)
}
\arguments{
\item{table}{a name of the modified table,}

\item{row_key}{a unique name identifing a row,}

\item{field}{names of modified columns,}

\item{value}{inserted values.}
}
\description{
The function updates a row in a selected table.
}
