% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_tab.R
\name{upload_table}
\alias{upload_table}
\title{Load table to the "geo_data" environment}
\usage{
upload_table(table, table_name)
}
\arguments{
\item{table}{a table to upload,}

\item{table_name}{a name of table in "geo_data" environment.}
}
\description{
The function loads a table created by a user
(e.g. with a "join_tab" function) to the "geo_data" environment.
}
